fix fastboot for use with module unification
