+++ lib/utilities/fastboot-app-module.js
@@ -7,16 +7,16 @@
 // The module defined here is prefixed with a `~` to make it less
 // likely to collide with user code, since it is not possible to
 // define a module with a name like this in the file system.
-module.exports = function fastbootAppModule(prefix, configAppAsString) {
+module.exports = function fastbootAppModule(prefix, moduleToRequire, configAppAsString) {
   return [
     "",
     "if (typeof FastBoot === 'undefined') {",
     "  if (!runningTests) {",
-    "    require('{{MODULE_PREFIX}}/app')['default'].create({{CONFIG_APP}});",
+    "    require('{{MODULE_PREFIX}}/{{MODULE}}')['default'].create({{CONFIG_APP}});",
     "  }",
     "}",
     "",
-    "define('~fastboot/app-factory', ['{{MODULE_PREFIX}}/app', '{{MODULE_PREFIX}}/config/environment'], function(App, config) {",
+    "define('~fastboot/app-factory', ['{{MODULE_PREFIX}}/{{MODULE}}', '{{MODULE_PREFIX}}/config/environment'], function(App, config) {",
     "  App = App['default'];",
     "  config = config['default'];",
     "",
@@ -27,5 +27,5 @@ module.exports = function fastbootAppModule(prefix, configAppAsString) {
     "  };",
     "});",
     ""
-  ].join("\n").replace(/\{\{MODULE_PREFIX\}\}/g, prefix).replace(/\{\{CONFIG_APP\}\}/g, configAppAsString);
+  ].join("\n").replace(/\{\{MODULE_PREFIX\}\}/g, prefix).replace(/\{\{MODULE\}\}/g, moduleToRequire).replace(/\{\{CONFIG_APP\}\}/g, configAppAsString);
 };
